refactor: Replace Volley by Picaso.
Square Picaso provides a transparent way of dealing with image caching,
eliminating the cumbersome LRU implementation and providing a more
readable code.

The material design color palette can be adjusted to the background
image, considering the its dominant, contrasting, and complementary
colors.

As soon as the image is downloaded by Picaso, those colors are
calculated by the command and new primary and accent colors are
generated. These are more harmonic with the background image.