feat: Create OnItemSelectedListener.
The RecyclerView does not have an setOnItemSelectedListener, so we need
to provide a OnItemSelectedListener to emulate the card selection.

The ArticleAdapter will implement the OnClickListener and associated
with the ViewHolder. When a card is clicked, it internally invokes the
OnItemSelectedListener interface method, that is implemented by the
ArticleListActivity and linked through a setter.